[FD20-445] Relocate check for $modified_hours
Remove the check for $modified_hours from the logic for displaying the modified hours block as a whole. Display the block even if there is no "Modified Hours" repeater value.

We often have content set that has no values in that repeater. Removing the repeater row requirement just cleaned up the back end.

But add the check for $modified_hours before the relevant foreach.