[FD20-499] Adjust input label
Several providers and their editors see patients via appointments, but those patients cannot make an appointment with that specific provider. An appointment can be made for the patient, and that provider can be assigned to that patient. So this phrasing will accomplish our goal of displaying information about providers for patients that have (or have had) an appointment with the provider without getting lost in the mechanics of how an appointment is made.