[FD20-3311] Change data source of clinical title
Clinical Title taxonomy has been renamed as Clinical Specialty. Clinical Specialty has a new input for Clinical Occupational Title. If that field has a value, set the clinical occupational title. If it does not have a value (as it likely won't immediately after this feature goes live), get the name of the taxonomy term instead.