[FD20-3311] Move 'mainContentOfPage' prop out of common schema
Formerly used the 'main' element as the value of the 'cssSelector' property.

Google structured data documentation says to "only use .class selectors for the cssSelector property."

The <main> element had the .content class, but that class is likely to be used in other elements as well. Needed a more unique class name to use for the 'cssSelector' property value.