[FD20-3311] Adjust $schema_type_list arg to allow null
uamswp_fad_schema_add_to_item_values

A value of null in this argument will still end the function immediately. But it will simplify the references to this function, eliminating the need to check for the $schema_type_list arg input value before every time the function is being called.